+++ extension/src/commands/CancelBuildCommand.ts
@@ -2,11 +2,11 @@ import * as vscode from "vscode";
 import { cancelBuild } from "../tasks/taskUtil";
 import { logger } from "../logger";
 import { Command } from "./Command";
-import { GradleClient } from "../client";
+import { TaskServerClient } from "../client";
 export const COMMAND_CANCEL_BUILD = "gradle.cancelBuild";
 
 export class CancelBuildCommand extends Command {
-    constructor(private client: GradleClient) {
+    constructor(private client: TaskServerClient) {
         super();
     }
     async run(cancellationKey: string, task?: vscode.Task): Promise<void> {