+++ CHANGELOG.md
@@ -4,6 +4,61 @@ All notable changes to the "vscode-gradle" extension will be documented in this 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 3.16.2
+### What's Changed
+* fix - Stop start build server when getRandomPipepath return empty by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1545
+* fix - Use 'java.import.gradle.java.home' when it's specified by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1552
+* fix - Inform that JDK 17 is required by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1549
+* fix - Refine findValidJavaHome logic by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1554
+
+## 3.16.1
+## What's Changed
+* enhancement - Support debug test delegation by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1536
+* fix - Gradle build server cannot be started when path has spaces by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1542
+* fix - Fail to start if import.gradle.java.home jdk version lower than 17 by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1547
+
+## 3.16.0
+## What's Changed
+* performance - Merge Task Server & Build Server by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1512
+* performance - Merge Language Server to Gradle Server by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1525
+* fix - Catch the error when running Gradle tests by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1524
+* fix - Wrong status code returned for test completion by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1532
+* fix - Override LanguageServer setTrace to avoid default Exception throw by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1533
+* fix - Get daemon status handle project folder name include empty space by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1537
+* documentation - Update architecture.md for merge server by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1520
+* documentation - Update doc after merge language server by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1528
+
+### 3.15.0
+## What's Changed
+* feat - Support delegate tests to Gradle build Server by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1510
+* fix - Only send request after initialization by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1503
+* fix - Add java 22 to compatibility matrix by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1505
+* refactor - Introduce gson to simplify the object parsing by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1509
+* fix - Slice the build target list to 1 when running gradle tests by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1518
+
+### 3.14.1
+### What's Changed
+* fix - Fix the importer version.
+
+## 3.14.0
+### What's Changed
+* enhancement - Bump Gradle Wrapper to v8.5 and change Java 21 min Gradle version by @JoseLion in https://github.com/microsoft/vscode-gradle/pull/1455
+* enhancement - Group the Gradle related menus in the file explorer into a submenu labeled 'Gradle' by @testforstephen in https://github.com/microsoft/vscode-gradle/pull/1474
+* enhancement - Upgrade Build Server for Gradle to 0.2.0
+* fix - Exclude the project itself becoming its project classpath entry by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1464
+* fix - Append the log message to build output if it's from compilation report. by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1490
+* fix - Set GradleExecution jdk use java.import.gradle.java.home by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1491
+* fix - Refresh the output folders after build tasks by @jdneo in https://github.com/microsoft/vscode-gradle/pull/1493
+* build - Fix buildJars task by @donat in https://github.com/microsoft/vscode-gradle/pull/1481
+* build - Add Build Server debug config by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1486
+* build - Move Gradle Daemons implementation from Java into Typescript by @Jiaaming in https://github.com/microsoft/vscode-gradle/pull/1489
+
+## New Contributors
+* @donat made their first contribution in https://github.com/microsoft/vscode-gradle/pull/1481
+* @Jiaaming made their first contribution in https://github.com/microsoft/vscode-gradle/pull/1486
+
+**Full Changelog**: https://github.com/microsoft/vscode-gradle/compare/3.13.5...3.14.0
+
 ## 3.13.5
 ### Added
 - Implement onWillUpdate() in GradleBuildServerBuildSupport. [PR#1405](https://github.com/microsoft/vscode-gradle/pull/1405)